Wrap all extension related information in 'extension_info'
