+++ server/webapp/WEB-INF/rails.new/app/presenters/api_v3/plugin/authentication_plugin_info_representer.rb
@@ -15,7 +15,7 @@
 ##########################################################################
 module ApiV3
   module Plugin
-    class AuthenticationPluginInfoRepresenter < BasePluginInfoRepresenter
+    class AuthenticationPluginInfoRepresenter < BaseRepresenter
       property :display_name
       property :display_image_url
