--- server/webapp/WEB-INF/rails.new/app/presenters/api_v3/plugin/config_repo_plugin_info_representer.rb
@@ -1,22 +1,0 @@
-##########################################################################
-# Copyright 2017 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-##########################################################################
-module ApiV3
-  module Plugin
-    class ConfigRepoPluginInfoRepresenter < BasePluginInfoRepresenter
-
-    end
-  end
-end
