+++ server/webapp/WEB-INF/rails.new/spec/presenters/api_v3/plugin/plugin_info_representer_spec.rb
@@ -16,18 +16,20 @@
 
 require 'spec_helper'
 
-describe ApiV3::Plugin::ConfigRepoPluginInfoRepresenter do
-  it 'should describe a notification plugin' do
+describe ApiV3::Plugin::PluginInfoRepresenter do
+
+  it 'should describe a NewPluginInfo object' do
     vendor = GoPluginDescriptor::Vendor.new('bob', 'https://bob.example.com')
     about = GoPluginDescriptor::About.new('Foo plugin', '1.2.3', '17.2.0', 'Does foo', vendor, ['Linux'])
-    descriptor = GoPluginDescriptor.new('foo.example', '1.0', about, nil, nil, false)
+    descriptor = GoPluginDescriptor.new('foo.example', '1.0', about, '/path/to/foo.jar', nil, false)
+    descriptor.markAsInvalid(%w(foo bar), java.lang.RuntimeException.new('boom!'))
 
-    config_repo_view = com.thoughtworks.go.plugin.domain.common.PluginView.new('role_config_view_template')
-    metadata = com.thoughtworks.go.plugin.domain.common.Metadata.new(true, false)
-    plugin_settings = com.thoughtworks.go.plugin.domain.common.PluggableInstanceSettings.new([com.thoughtworks.go.plugin.domain.common.PluginConfiguration.new('memberOf', metadata)], config_repo_view)
+    plugin_view = com.thoughtworks.go.plugin.domain.common.PluginView.new('plugin_view_template')
+    plugin_metadata = com.thoughtworks.go.plugin.domain.common.Metadata.new(true, false)
+    plugin_settings = com.thoughtworks.go.plugin.domain.common.PluggableInstanceSettings.new([com.thoughtworks.go.plugin.domain.common.PluginConfiguration.new('memberOf', plugin_metadata)], plugin_view)
 
-    plugin_info = com.thoughtworks.go.plugin.domain.configrepo.ConfigRepoPluginInfo.new(descriptor, plugin_settings)
-    actual_json = ApiV3::Plugin::ConfigRepoPluginInfoRepresenter.new(plugin_info).to_hash(url_builder: UrlBuilder.new)
+    plugin_info = com.thoughtworks.go.plugin.domain.common.PluginInfo.new(descriptor, 'plugin-type', plugin_settings)
+    actual_json = ApiV3::Plugin::PluginInfoRepresenter.new(plugin_info).to_hash(url_builder: UrlBuilder.new)
 
     expect(actual_json).to have_links(:self, :doc, :find)
     expect(actual_json).to have_link(:self).with_url('http://test.host/api/admin/plugin_info/foo.example')
@@ -37,12 +39,13 @@ describe ApiV3::Plugin::ConfigRepoPluginInfoRepresenter do
 
     expect(actual_json).to eq({
                                 id: 'foo.example',
-                                type: 'configrepo',
-                                status: {
-                                  state: 'active'
-                                },
-                                plugin_file_location: nil,
+                                type: 'plugin-type',
+                                plugin_file_location: '/path/to/foo.jar',
                                 bundled_plugin: false,
+                                status: {
+                                  state: 'invalid',
+                                  messages: %w(foo bar)
+                                },
                                 about: {
                                   name: 'Foo plugin',
                                   version: '1.2.3',
@@ -53,8 +56,20 @@ describe ApiV3::Plugin::ConfigRepoPluginInfoRepresenter do
                                     name: 'bob',
                                     url: 'https://bob.example.com'}
                                 },
-                                plugin_settings: ApiV3::Plugin::PluggableInstanceSettingsRepresenter.new(plugin_settings).to_hash(url_builder: UrlBuilder.new)
+                                plugin_settings: {
+                                  configurations: [
+                                    {
+                                      key: "memberOf",
+                                      metadata: {
+                                        secure: false,
+                                        required: true
+                                      }
+                                    }
+                                  ],
+                                  view: {
+                                    template: "plugin_view_template"
+                                  }
+                                }
                               })
-
   end
 end
