+++ server/webapp/WEB-INF/rails.new/spec/presenters/api_v3/plugin/scm_plugin_info_representer_spec.rb
@@ -29,32 +29,13 @@ describe ApiV3::Plugin::SCMPluginInfoRepresenter do
     plugin_info = com.thoughtworks.go.plugin.domain.scm.SCMPluginInfo.new(descriptor, 'Foo task', scm_settings, nil)
     actual_json = ApiV3::Plugin::SCMPluginInfoRepresenter.new(plugin_info).to_hash(url_builder: UrlBuilder.new)
 
-    expect(actual_json).to have_links(:self, :doc, :find)
-    expect(actual_json).to have_link(:self).with_url('http://test.host/api/admin/plugin_info/foo.example')
-    expect(actual_json).to have_link(:doc).with_url('https://api.gocd.org/#plugin-info')
-    expect(actual_json).to have_link(:find).with_url('http://test.host/api/admin/plugin_info/:plugin_id')
+    expect(actual_json).to have_links(:scm_settings_doc)
+    expect(actual_json).to have_link(:scm_settings_doc).with_url('https://api.gocd.org/#scms')
     actual_json.delete(:_links)
 
     expect(actual_json).to eq({
-                                id: 'foo.example',
-                                type: 'scm',
-                                status: {
-                                  state: 'active'
-                                },
-                                plugin_file_location: nil,
-                                bundled_plugin: false,
-                                about: {
-                                  name: 'Foo plugin',
-                                  version: '1.2.3',
-                                  target_go_version: '17.2.0',
-                                  description: 'Does foo',
-                                  target_operating_systems: ['Linux'],
-                                  vendor: {
-                                    name: 'bob',
-                                    url: 'https://bob.example.com'}
-                                },
                                 display_name: 'Foo task',
-                                extension_settings: {scm_settings: ApiV3::Plugin::PluggableInstanceSettingsRepresenter.new(scm_settings).to_hash(url_builder: UrlBuilder.new)},
+                                scm_settings: ApiV3::Plugin::PluggableInstanceSettingsRepresenter.new(scm_settings).to_hash(url_builder: UrlBuilder.new),
                               })
 
   end
